// lib/types/components/progress.d.ts
import { type ReactNode } from "react";
import { type TColors } from "../generic";

export type TProgressVariant = "filled" | "outlined" | "text";
export type TProgressSize = "sm" | "md" | "lg";
export type TProgressColors = TColors;
export type TProgressLabel = ReactNode;
export type TProgressValue = number;
export type TProgressBar = number;
export type TProgressBarProps = object;
export type TProgressClassname = string;